+++ packages/vkui/src/components/SplitCol/SplitColContext.tsx
@@ -1,0 +1,13 @@
+import * as React from 'react';
+
+export interface SplitColContextProps {
+  colRef: React.RefObject<HTMLDivElement> | null;
+  animate: boolean;
+}
+
+export const SplitColContext = React.createContext<SplitColContextProps>({
+  colRef: null,
+  animate: true,
+});
+
+export const useSplitCol = () => React.useContext(SplitColContext);
